@@ -44,7 +44,7 @@ def getCheckedOutGitCommitHash() {
 group = 'fr.olympa'
 version = '0.0.1-SNAPSHOT'
 description = 'olympa-api'
-java.sourceCompatibility = JavaVersion.VERSION_11
+java.sourceCompatibility = JavaVersion.VERSION_16
 
 def fullVersion = description + '-' + version + '-' + getCheckedOutGitCommitHash()
 
